Crash scenario for aws
Crash scenario with one machine for aws.

+++ tests/sles4sap/crash/destroy.pm
@@ -4,7 +4,7 @@
 # SPDX-License-Identifier: FSFAP
 # Maintainer: QE-SAP <qe-sap@suse.de>
 # Summary: Public Cloud - Resource Cleanup
-# This module deletes the Azure resource group to:
+# This module deletes resources from cloud:
 # - Free up cloud resources
 # - Avoid additional cost
 # - Clean up the environment
@@ -15,20 +15,49 @@ use Mojo::Base 'publiccloud::basetest';
 use testapi;
 use mmapi 'get_current_job_id';
 use serial_terminal 'select_serial_terminal';
+use sles4sap::aws_cli;
 
 
 sub run {
     my ($self) = @_;
 
-    die('Azure is the only CSP supported for the moment')
-      unless check_var('PUBLIC_CLOUD_PROVIDER', 'AZURE');
+    die('CSP not supported for the moment')
+      unless check_var('PUBLIC_CLOUD_PROVIDER', get_required_var('PUBLIC_CLOUD_PROVIDER'));
 
     select_serial_terminal;
+    my $aws_prefix = get_var('DEPLOY_PREFIX', 'clne');
+    my $job_id = $aws_prefix . get_current_job_id();
 
-    my $rg = get_required_var('RG');
-    record_info('AZURE CLEANUP', "Deleting resource group: $rg");
-    assert_script_run("az group delete --name $rg -y", timeout => 600);
-    assert_script_run("az group wait --name $rg --deleted");
+    my $provider_cloud_name = get_required_var('PUBLIC_CLOUD_PROVIDER');
+    if ($provider_cloud_name eq 'AZURE') {
+        my $rg = get_required_var('RG');
+        record_info('AZURE CLEANUP', "Deleting resource group: $rg");
+        assert_script_run("az group delete --name $rg -y", timeout => 600);
+        assert_script_run("az group wait --name $rg --deleted");
+    }
+    if ($provider_cloud_name eq 'EC2') {
+        my $region = get_required_var('PUBLIC_CLOUD_REGION');
+        my $instance_id = aws_get_vm_id($region, $job_id);
+
+        my $vpc_id = aws_get_vpc_id($region, $job_id);
+        # Terminate instance and wait
+        script_run("aws ec2 terminate-instances --instance-ids $instance_id --region $region");
+        script_run("aws ec2 wait instance-terminated --instance-ids $instance_id --region $region", timeout => 300);
+        # Delete all resources
+        assert_script_run("aws ec2 delete-security-group --group-id " . aws_get_security_group_id($region, $job_id) . " --region $region");
+        assert_script_run("aws ec2 delete-subnet --subnet-id " . aws_get_subnet_id($region, $job_id) . " --region $region");
+        my $igw_id = aws_get_internet_gateway_id($region, $job_id);
+        script_run("aws ec2 detach-internet-gateway --vpc-id $vpc_id --internet-gateway-id $igw_id --region $region");
+        script_run("aws ec2 delete-internet-gateway --internet-gateway-id $igw_id --region $region");
+        my $rtb_ids = script_output(
+            "aws ec2 describe-route-tables --filters Name=vpc-id,Values=$vpc_id" .
+              " --query 'RouteTables[?Associations[0].Main!=\`true\`].RouteTableId'" .
+              " --output text --region $region");
+        assert_script_run("aws ec2 delete-route-table --route-table-id $_ --region $region") for split(/\s+/, $rtb_ids);
+
+        # Delete everything else (AWS handles dependencies automatically if we wait)
+        script_run("aws ec2 delete-vpc --vpc-id $vpc_id --region $region");
+    }
     record_info('Done', 'Test finished');
 }
